Refactored app.js code
