+++ app.js
@@ -3,6 +3,8 @@ const express = require('express');
 const path = require('path');
 const http = require('http');
 const ConstReqParser = require('./app/customReqParser.class').CustomReqParser;
+const errorHandler = require('./app/errorHandler').errorHandler;
+const findAndSendWeather = require('./app/findAndSendWeather').findAndSendWeather;
 
 const myapp = express();
 const port = (process.env.PORT || 5000);
@@ -12,12 +14,6 @@ myapp.set('port', port);
 myapp.set('views', './views');
 myapp.set('view engine', 'pug');
 myapp.set('trust proxy', true);
-
-function errorHandler(error) {
-  console.log(`error message: ${error}`);
-  return { error: `encountered server ${error}` };
-}
-
 // Routing
 
 myapp.get('/', (req, res) => {
@@ -53,7 +49,7 @@ myapp.get('/dameelclima', (req, res) => {
       hostname: 'ip-api.com',
       path: `/json/${ipaddress}`,
     };
-    const _serverReq = http.get(_options, (response) => {
+    http.get(_options, (response) => {
       try {
         if (response.statusCode !== 200) {
           throw new Error('response with wrong status code');
@@ -63,7 +59,7 @@ myapp.get('/dameelclima', (req, res) => {
         response.setEncoding('utf8');
         let rawData = '';
         let i = 1;
-        response.on('data', (chunk)=> {
+        response.on('data', (chunk) => {
           console.log(`data received ${i} times`);
           rawData += chunk;
           i += 1;
@@ -90,56 +86,7 @@ myapp.get('/dameelclima', (req, res) => {
   }
 });
 
-function findAndSendWeather(locationData, res){
-  const permissibleUrls = [/https:\/\/s\.codepen\.io\/?/g, /https:\/\/noce2\.github\.io\/?/g];
-  const reqOrigin = req.get('Origin');
-  const detectedUrl = permissibleUrls.filter(each => each.test(reqOrigin));
-  if (detectedUrl.length === 1) {
-    res.set({
-      'Access-Control-Allow-Origin': reqOrigin,
-    });
-  }
-  const apiTarget = 'api.openweathermap.org';
-  const queryParams = {
-    lat: locationData.lat,
-    lon: locationData.lon,
-    APPID: '078ecb2576afa59e6e132d1ce4c68684',
-    units: 'metric',
-  };
-  if (queryParams.lat && queryParams.lon && queryParams.APPID && queryParams.units) {
-    const querystring = `/data/2.5/weather?lat=${queryParams.lat}&lon=${queryParams.lon}&APPID=${queryParams.APPID}&units=${queryParams.units}`;
-    const options = {
-      hostname: apiTarget,
-      path: querystring,
-    };
-    http.get(options, (_res) => {
-      let receivedData = '';
-      let receivedTimes = 1;
-      try {
-        if (_res.statusCode !== 200) {
-          throw new Error('response with wrong status code');
-        } else if (!(/application\/json/).test(_res.headers['content-type'])) {
-          throw new Error('response with wrong data type');
-        }
-        _res.setEncoding('utf8');
-        _res.on('data', (chunk) => {
-          console.log(`data received ${receivedTimes} times`);
-          receivedData += chunk;
-          receivedTimes += 1;
-        });
-        _res.on('end', () => {
-          res.send(receivedData);
-        });
-      } catch (_err) {
-        const message = errorHandler(_err);
-        console.log(message);
-        res.send(message);
-      }
-    });
-  } else {
-    res.send(JSON.stringify({ error: 'the necessary query parameters were not attached to the request' }));
-  }
-}
+
 
 // Starting the application
 myapp.listen(myapp.get('port'), () => {
